% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OR_95CI.R
\name{OR_95CI}
\alias{OR_95CI}
\title{odds ratio with 95 percent confidence interval}
\usage{
OR_95CI(coef, se, siglevel, roundto)
}
\arguments{
\item{coef}{beta coefficient from logistic regression}

\item{se}{standard error of the beta coefficient}

\item{siglevel}{alpha/significance level you wish to use (ex. 0.95)}

\item{roundto}{the multiple to which the number will be rounded (default is 1)}
}
\value{
an odds ratio with an upper and lower 95 percent confidence interval
}
\description{
a function that returns the odds ratio with a 95 percent confidence interval
}
\examples{
library(rms)
toy.lrm <- rms::lrm(y ~ x1 + x2, data = toydata, x = TRUE, y = TRUE)
toy.lrm
OR_95CI(1.0055, 0.2534, 0.95, 1)
OR_95CI(0.8937, 0.5948, 0.95, 2)
}
\author{
Kayla Esser
}
