% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toydata.R
\docType{data}
\name{toydata}
\alias{toydata}
\title{Toy Data}
\format{
a data frame with 100 rows and 3 variables
\describe{
\item{x1}{a continuous predicter}
\item{x2}{a binary predictor}
\item{y}{a binary outcome}
}
}
\usage{
toydata
}
\description{
Test data from directed reading course at U of T
}
\keyword{datasets}
